newsapp/newsportal/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-allauth==0.63.3
django-apscheduler==0.6.2
django-filter==24.2
idna==3.7
kombu==5.4.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
